+++ backend/common/types.ts
@@ -38,3 +38,18 @@ export interface QuizState {
     isActive : boolean;
     quizData : QuizItem | null;
 }
+
+type MessageType =
+  | "correct"
+  | "voteStart"
+  | "voteEnd"
+  | "warning"
+  | "quizStart"
+  | "quizEnd";
+
+export interface SystemMessageProps {
+  type: MessageType;
+  nickName?: string;
+  time: string;
+  items?: VoteItem[];
+}
