+++ backend/socket/messageHandler.ts
@@ -1,7 +1,7 @@
 import pool from '../db/mariadb';
 import { ResultSetHeader } from 'mysql2/promise';
 import { Server, Socket } from 'socket.io';
-import type { MessagePayload } from '../common/types';
+import type { MessagePayload, SystemMessageProps } from '../common/types';
 
 const saveSendMessageToDB = async (userId: number, nickname: string, content: string): Promise<ResultSetHeader | null> => {
     const query = 'INSERT INTO messages (user_id, nickname, content, created_at) VALUES (?, ?, ?, NOW())';
@@ -49,5 +49,9 @@ export function handleMessage(io: Server, socket: Socket) {
             console.error('DB : 메시지 처리 중 오류 발생', err);
             // socket.emit('SEND_MESSAGE_ERROR', {message: '메시지 전송 중 오류가 발생했습니다.'});
         }
+    });
+
+    socket.on('SEND_SYSTEM_MSG', async (data : SystemMessageProps) => {
+        io.emit('RECEIVE_SYSTEM_MSG', data);
     });
 }
