+++ frontend/src/socket/messageHandler.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { socket } from './socketManager';
-import type { MessagePayload } from '../common/types';
+import type { MessagePayload, SystemMessageProps } from '../common/types';
 
 export const useMessageHandler = () => {
     const [message, setMessage] = useState('');
@@ -16,8 +16,15 @@ export const useMessageHandler = () => {
             setMessages(prevMessages => [...prevMessages, msg]); 
         });
 
+        socket.on('RECEIVE_SYSTEM_MSG', (msg: SystemMessageProps) => {
+            console.log('서버로부터 시스템 메시지 수신:', msg);
+            // 프론트 TODO : 모든 사용자에게 시스템 메세지 표시
+            // 기존에 setSystemMessages를 여기서 실행해야 투표/퀴즈 시작 당사자에게 중복으로 출력되지 않음
+        });
+
         return () => {
             socket.off('RECEIVE_MESSAGE');
+            socket.off('RECEIVE_SYSTEM_MSG');
         };
     }, []);
